python/requirements.txt
pandas==1.3.5
streamlit==1.34.0
mysql-connector-python